Extract ProjectionHelper & ViewportHelper
